Implemented compute_reservation_create_for_existing_instance_template, created tests
