--- compute/cloud-client/src/test/java/compute/reservation/CreateReservationWithRegionalInstanceTemplateIT.java
@@ -1,224 +1,0 @@
-/*
- * Copyright 2024 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package compute.reservation;
-
-import static com.google.common.truth.Truth.assertThat;
-import static com.google.common.truth.Truth.assertWithMessage;
-
-import com.google.cloud.compute.v1.AttachedDisk;
-import com.google.cloud.compute.v1.AttachedDiskInitializeParams;
-import com.google.cloud.compute.v1.DeleteRegionInstanceTemplateRequest;
-import com.google.cloud.compute.v1.InsertRegionInstanceTemplateRequest;
-import com.google.cloud.compute.v1.InstanceProperties;
-import com.google.cloud.compute.v1.InstanceTemplate;
-import com.google.cloud.compute.v1.NetworkInterface;
-import com.google.cloud.compute.v1.Operation;
-import com.google.cloud.compute.v1.RegionInstanceTemplatesClient;
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.PrintStream;
-import java.util.UUID;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.Timeout;
-import org.junit.runner.RunWith;
-import org.junit.runners.JUnit4;
-
-@RunWith(JUnit4.class)
-@Timeout(value = 10, unit = TimeUnit.MINUTES)
-class CreateReservationWithRegionalInstanceTemplateIT {
-  private static final String PROJECT_ID = System.getenv("GOOGLE_CLOUD_PROJECT");
-  private static final String DEFAULT_ZONE = "us-central1-a";
-  private static final String INSTANCE_TEMPLATE_NAME = "test-instance-" + UUID.randomUUID();
-  private static String INSTANCE_TEMPLATE_URI;
-  private static String RESERVATION_NAME;
-  private static final int NUMBER_OF_VMS = 3;
-  private static final boolean SPECIFIC_RESERVATION_REQUIRED = true;
-
-  private ByteArrayOutputStream stdOut;
-
-  // Check if the required environment variables are set.
-  public static void requireEnvVar(String envVarName) {
-    assertWithMessage(String.format("Missing environment variable '%s' ", envVarName))
-        .that(System.getenv(envVarName)).isNotEmpty();
-  }
-
-  @BeforeAll
-  public static void setup()
-      throws IOException, ExecutionException, InterruptedException, TimeoutException {
-    final PrintStream out = System.out;
-    ByteArrayOutputStream stdOut = new ByteArrayOutputStream();
-    System.setOut(new PrintStream(stdOut));
-    requireEnvVar("GOOGLE_APPLICATION_CREDENTIALS");
-    requireEnvVar("GOOGLE_CLOUD_PROJECT");
-    INSTANCE_TEMPLATE_URI = String.format("projects/%s/regions/us-central1/instanceTemplates/%s",
-        PROJECT_ID, INSTANCE_TEMPLATE_NAME);
-    RESERVATION_NAME = "test-reserv-" + UUID.randomUUID();
-
-    // Create templates.
-    CreateReservationWithRegionalInstanceTemplateIT.createInstanceTemplate(
-        PROJECT_ID, INSTANCE_TEMPLATE_NAME, DEFAULT_ZONE);
-    assertThat(stdOut.toString()).contains("Instance Template Operation Status: DONE");
-
-    stdOut.close();
-    System.setOut(out);
-  }
-
-  @AfterAll
-  public static void cleanup()
-      throws IOException, ExecutionException, InterruptedException, TimeoutException {
-    final PrintStream out = System.out;
-    ByteArrayOutputStream stdOut = new ByteArrayOutputStream();
-    System.setOut(new PrintStream(stdOut));
-
-    // Delete the regional instance template
-    CreateReservationWithRegionalInstanceTemplateIT.deleteRegionalInstanceTemplate(
-        PROJECT_ID, DEFAULT_ZONE, INSTANCE_TEMPLATE_NAME);
-    assertThat(stdOut.toString())
-        .contains("Instance template deletion operation status for " + INSTANCE_TEMPLATE_NAME);
-
-//    // Delete reservation
-//    DeleteReservation.deleteReservationWithRegionalLocation(
-//        PROJECT_ID, DEFAULT_ZONE, RESERVATION_NAME);
-//    assertThat(stdOut.toString()).contains("Deleted reservation: " + RESERVATION_NAME);
-
-    stdOut.close();
-    System.setOut(out);
-  }
-
-  @BeforeEach
-  public void beforeEach() {
-    stdOut = new ByteArrayOutputStream();
-    System.setOut(new PrintStream(stdOut));
-  }
-
-  @AfterEach
-  public void afterEach() {
-    stdOut = null;
-    System.setOut(null);
-  }
-
-  @Test
-  public void testCrateReservationWithRegionInstanceTemplate()
-      throws IOException, ExecutionException, InterruptedException, TimeoutException {
-    CreateReservationForInstanceTemplate.createReservationForInstanceTemplate(
-        PROJECT_ID, RESERVATION_NAME, INSTANCE_TEMPLATE_URI,
-        NUMBER_OF_VMS, DEFAULT_ZONE, SPECIFIC_RESERVATION_REQUIRED);
-
-    assertThat(stdOut.toString()).contains("Reservation created. Operation Status: DONE");
-  }
-
-  // Creates a new instance template in the specified project and region.
-  public static void createInstanceTemplate(
-      String projectId, String templateName, String zone)
-      throws IOException, ExecutionException, InterruptedException, TimeoutException {
-    try (RegionInstanceTemplatesClient templatesClientRegion =
-             RegionInstanceTemplatesClient.create()) {
-
-      String machineType = "n1-standard-1"; // Example machine type
-      String sourceImage = "projects/debian-cloud/global/images/family/debian-11"; // Example image
-      String region = zone.substring(0, zone.lastIndexOf('-')); // Extract the region from the zone
-
-      // Define the boot disk for the instance template
-      AttachedDisk attachedDisk = AttachedDisk.newBuilder()
-          .setInitializeParams(AttachedDiskInitializeParams.newBuilder()
-              .setSourceImage(sourceImage)
-              .setDiskType("pd-balanced") // Example disk type
-              .setDiskSizeGb(100L) // Example disk size
-              .build())
-          .setAutoDelete(true)
-          .setBoot(true)
-          .build();
-
-      // Define the network interface for the instance template
-      // Note: The subnetwork must be in the same region as the instance template.
-      NetworkInterface networkInterface = NetworkInterface.newBuilder()
-          .setName("my-network-test")
-          .setSubnetwork(String.format("projects/%s/regions/%s/subnetworks/default",
-              PROJECT_ID, region))
-          .build();
-
-      // Define the instance properties for the template
-      InstanceProperties instanceProperties = InstanceProperties.newBuilder()
-          .addDisks(attachedDisk)
-          .setMachineType(machineType)
-          .addNetworkInterfaces(networkInterface)
-          .build();
-
-      // Build the instance template object
-      InstanceTemplate instanceTemplate = InstanceTemplate.newBuilder()
-          .setName(templateName)
-          .setProperties(instanceProperties)
-          .build();
-
-      // Create the request to insert the instance template
-      InsertRegionInstanceTemplateRequest insertInstanceTemplateRequest =
-          InsertRegionInstanceTemplateRequest
-              .newBuilder()
-              .setProject(projectId)
-              .setRegion(region)
-              .setInstanceTemplateResource(instanceTemplate)
-              .build();
-
-      // Send the request and wait for the operation to complete
-      Operation response = templatesClientRegion.insertAsync(insertInstanceTemplateRequest)
-          .get(3, TimeUnit.MINUTES);
-
-      if (response.hasError()) {
-        System.out.println("Instance Template creation failed! " + response);
-        return;
-      }
-      System.out.printf("Instance Template Operation Status: %s%n", response.getStatus());
-    }
-  }
-
-
-  // Delete a regional instance template.
-  private static void deleteRegionalInstanceTemplate(
-      String projectId, String zone, String templateName)
-      throws IOException, ExecutionException, InterruptedException, TimeoutException {
-    try (RegionInstanceTemplatesClient regionInstanceTemplatesClient =
-             RegionInstanceTemplatesClient.create()) {
-      String region = zone.substring(0, zone.lastIndexOf('-')); // Extract the region from the zone
-
-      DeleteRegionInstanceTemplateRequest deleteInstanceTemplateRequest =
-          DeleteRegionInstanceTemplateRequest
-              .newBuilder()
-              .setProject(projectId)
-              .setRegion(region)
-              .setInstanceTemplate(templateName)
-              .build();
-
-      Operation response = regionInstanceTemplatesClient.deleteAsync(
-          deleteInstanceTemplateRequest).get(3, TimeUnit.MINUTES);
-
-      if (response.hasError()) {
-        System.out.println("Instance template deletion failed ! ! " + response);
-        return;
-      }
-      System.out.printf("Instance template deletion operation status for %s: %s ", templateName,
-          response.getStatus());
-    }
-  }
-}
-
