+++ compute/cloud-client/src/test/java/compute/reservation/ReservationIT.java
@@ -17,14 +17,21 @@
 package compute.reservation;
 
 import static com.google.common.truth.Truth.assertThat;
-import static compute.Util.getZone;
+import static com.google.common.truth.Truth.assertWithMessage;
 
-import com.google.cloud.compute.v1.AllocationSpecificSKUAllocationReservedInstanceProperties;
-import com.google.cloud.compute.v1.AllocationSpecificSKUReservation;
-import com.google.cloud.compute.v1.InsertReservationRequest;
+import com.google.cloud.compute.v1.AttachedDisk;
+import com.google.cloud.compute.v1.AttachedDiskInitializeParams;
+import com.google.cloud.compute.v1.DeleteRegionInstanceTemplateRequest;
+import com.google.cloud.compute.v1.InsertRegionInstanceTemplateRequest;
+import com.google.cloud.compute.v1.InstanceProperties;
+import com.google.cloud.compute.v1.InstanceTemplate;
+import com.google.cloud.compute.v1.NetworkInterface;
 import com.google.cloud.compute.v1.Operation;
+import com.google.cloud.compute.v1.RegionInstanceTemplatesClient;
 import com.google.cloud.compute.v1.Reservation;
 import com.google.cloud.compute.v1.ReservationsClient;
+import compute.CreateInstanceTemplate;
+import compute.DeleteInstanceTemplate;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.PrintStream;
@@ -32,58 +39,98 @@ import java.util.UUID;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
+import org.junit.Assert;
+import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.Timeout;
+import org.junit.runner.RunWith;
+import org.junit.runners.JUnit4;
 
+@RunWith(JUnit4.class)
+@Timeout(value = 10, unit = TimeUnit.MINUTES)
 public class ReservationIT {
 
-  private static String PROJECT_ID;
-  private static String ZONE;
-  private static String RESERVATION_NAME;
+  private static final String PROJECT_ID = System.getenv("GOOGLE_CLOUD_PROJECT");
+  private static final String ZONE = "us-central1-a";
+  private static String RESERVATION_NAME_1;
+  private static String RESERVATION_NAME_2;
+  private static final String GLOBAL_INSTANCE_TEMPLATE_NAME =
+      "test-global-instance-" + UUID.randomUUID();
+  private static final String REGIONAL_INSTANCE_TEMPLATE_NAME =
+      "test-regional-instance-" + UUID.randomUUID();
+  private static String GLOBAL_INSTANCE_TEMPLATE_URI;
+  private static String REGIONAL_INSTANCE_TEMPLATE_URI;
+  private static final int NUMBER_OF_VMS = 3;
+  private static final boolean SPECIFIC_RESERVATION_REQUIRED = true;
 
   private ByteArrayOutputStream stdOut;
+
+  // Check if the required environment variables are set.
+  public static void requireEnvVar(String envVarName) {
+    assertWithMessage(String.format("Missing environment variable '%s' ", envVarName))
+        .that(System.getenv(envVarName)).isNotEmpty();
+  }
 
   @BeforeAll
   public static void setUp()
       throws IOException, ExecutionException, InterruptedException, TimeoutException {
-    PROJECT_ID = System.getenv("GOOGLE_CLOUD_PROJECT");
-    ZONE = getZone();
-    RESERVATION_NAME = "test-reservation-" + UUID.randomUUID();
+    final PrintStream out = System.out;
+    ByteArrayOutputStream stdOut = new ByteArrayOutputStream();
+    System.setOut(new PrintStream(stdOut));
+    requireEnvVar("GOOGLE_APPLICATION_CREDENTIALS");
+    requireEnvVar("GOOGLE_CLOUD_PROJECT");
+    RESERVATION_NAME_1 = "test-reservation1-" + UUID.randomUUID();
+    RESERVATION_NAME_2 = "test-reservation2-" + UUID.randomUUID();
+    GLOBAL_INSTANCE_TEMPLATE_URI = String.format("projects/%s/global/instanceTemplates/%s",
+        PROJECT_ID, GLOBAL_INSTANCE_TEMPLATE_NAME);
+    REGIONAL_INSTANCE_TEMPLATE_URI =
+        String.format("projects/%s/regions/us-central1/instanceTemplates/%s",
+        PROJECT_ID, REGIONAL_INSTANCE_TEMPLATE_NAME);
 
-    // Create the reservation.
-    try (ReservationsClient reservationsClient = ReservationsClient.create()) {
+    // Create instance template with GLOBAL location.
+    CreateInstanceTemplate.createInstanceTemplate(PROJECT_ID, GLOBAL_INSTANCE_TEMPLATE_NAME);
+    assertThat(stdOut.toString())
+        .contains("Instance Template Operation Status " + GLOBAL_INSTANCE_TEMPLATE_NAME);
+    // Create instance template with REGIONAL location.
+    ReservationIT.createRegionalInstanceTemplate(
+        PROJECT_ID, REGIONAL_INSTANCE_TEMPLATE_NAME, ZONE);
+    assertThat(stdOut.toString()).contains("Instance Template Operation Status: DONE");
 
-      Reservation reservation = Reservation.newBuilder()
-          .setName(RESERVATION_NAME)
-          .setSpecificReservation(
-              AllocationSpecificSKUReservation.newBuilder()
-                  .setCount(1)
-                  .setInstanceProperties(
-                      AllocationSpecificSKUAllocationReservedInstanceProperties.newBuilder()
-                          .setMachineType("n1-standard-1")
-                          .build())
-                  .build())
-          .build();
+    stdOut.close();
+    System.setOut(out);
+  }
 
-      InsertReservationRequest reservationRequest = InsertReservationRequest.newBuilder()
-          .setProject(PROJECT_ID)
-          .setZone(ZONE)
-          .setReservationResource(reservation)
-          .build();
+  @AfterAll
+  public static void cleanup()
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
+    final PrintStream out = System.out;
+    ByteArrayOutputStream stdOut = new ByteArrayOutputStream();
+    System.setOut(new PrintStream(stdOut));
 
-      Operation response = reservationsClient.insertAsync(reservationRequest)
-          .get(3, TimeUnit.MINUTES);
+    // Delete instance template with GLOBAL location.
+    DeleteInstanceTemplate.deleteInstanceTemplate(PROJECT_ID, GLOBAL_INSTANCE_TEMPLATE_NAME);
+    assertThat(stdOut.toString())
+        .contains("Instance template deletion operation status for "
+            + GLOBAL_INSTANCE_TEMPLATE_NAME);
 
-      if (response.getStatus() == Operation.Status.DONE) {
-        System.out.println("Reservation created.");
-      } else {
-        System.out.println("Reservation creation failed!");
-      }
+    // Delete instance template with REGIONAL location.
+    ReservationIT.deleteRegionalInstanceTemplate(
+        PROJECT_ID, ZONE, REGIONAL_INSTANCE_TEMPLATE_NAME);
+    assertThat(stdOut.toString())
+        .contains("Instance template deletion operation status for "
+            + REGIONAL_INSTANCE_TEMPLATE_NAME);
 
-      assertThat(reservation.getName()).isEqualTo(RESERVATION_NAME);
-    }
+    // Verify reservations are deleted
+    DeleteReservation.deleteReservation(PROJECT_ID, ZONE, RESERVATION_NAME_1);
+    assertThat(stdOut.toString()).contains("Deleted reservation: " + RESERVATION_NAME_1);
+    DeleteReservation.deleteReservation(PROJECT_ID, ZONE, RESERVATION_NAME_2);
+    assertThat(stdOut.toString()).contains("Deleted reservation: " + RESERVATION_NAME_2);
+
+    stdOut.close();
+    System.setOut(out);
   }
 
   @BeforeEach
@@ -99,10 +146,128 @@ public class ReservationIT {
   }
 
   @Test
-  public void testDeleteReservation()
+  public void testCrateReservationWithGlobalInstanceTemplate()
       throws IOException, ExecutionException, InterruptedException, TimeoutException {
-    DeleteReservation.deleteReservation(PROJECT_ID, ZONE, RESERVATION_NAME);
+    CreateReservationForInstanceTemplate.createReservationForInstanceTemplate(
+        PROJECT_ID, RESERVATION_NAME_1,
+        GLOBAL_INSTANCE_TEMPLATE_URI, NUMBER_OF_VMS, ZONE,
+        SPECIFIC_RESERVATION_REQUIRED);
 
-    assertThat(stdOut.toString()).contains("Deleted reservation: " + RESERVATION_NAME);
+    try (ReservationsClient reservationsClient = ReservationsClient.create()) {
+      Reservation reservation = reservationsClient.get(PROJECT_ID, ZONE, RESERVATION_NAME_1);
+
+      assertThat(stdOut.toString()).contains("Reservation created. Operation Status: DONE");
+      Assert.assertEquals(NUMBER_OF_VMS,
+          reservation.getSpecificReservation().getCount());
+      Assert.assertTrue(reservation.getSpecificReservation()
+          .getSourceInstanceTemplate().contains(GLOBAL_INSTANCE_TEMPLATE_NAME));
+      Assert.assertTrue(reservation.getZone().contains(ZONE));
+      Assert.assertEquals(RESERVATION_NAME_1, reservation.getName());
+      Assert.assertEquals(SPECIFIC_RESERVATION_REQUIRED,
+          reservation.getSpecificReservationRequired());
+    }
+  }
+
+  @Test
+  public void testCrateReservationWithRegionInstanceTemplate()
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
+    CreateReservationForInstanceTemplate.createReservationForInstanceTemplate(
+        PROJECT_ID, RESERVATION_NAME_2, REGIONAL_INSTANCE_TEMPLATE_URI,
+        NUMBER_OF_VMS, ZONE, SPECIFIC_RESERVATION_REQUIRED);
+
+    assertThat(stdOut.toString()).contains("Reservation created. Operation Status: DONE");
+  }
+
+  // Creates a new instance template with the REGIONAL location.
+  public static void createRegionalInstanceTemplate(
+      String projectId, String templateName, String zone)
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
+    try (RegionInstanceTemplatesClient templatesClientRegion =
+             RegionInstanceTemplatesClient.create()) {
+
+      String machineType = "n1-standard-1"; // Example machine type
+      String sourceImage = "projects/debian-cloud/global/images/family/debian-11"; // Example image
+      String region = zone.substring(0, zone.lastIndexOf('-')); // Extract the region from the zone
+
+      // Define the boot disk for the instance template
+      AttachedDisk attachedDisk = AttachedDisk.newBuilder()
+          .setInitializeParams(AttachedDiskInitializeParams.newBuilder()
+              .setSourceImage(sourceImage)
+              .setDiskType("pd-balanced") // Example disk type
+              .setDiskSizeGb(100L) // Example disk size
+              .build())
+          .setAutoDelete(true)
+          .setBoot(true)
+          .build();
+
+      // Define the network interface for the instance template
+      // Note: The subnetwork must be in the same region as the instance template.
+      NetworkInterface networkInterface = NetworkInterface.newBuilder()
+          .setName("my-network-test")
+          .setSubnetwork(String.format("projects/%s/regions/%s/subnetworks/default",
+              PROJECT_ID, region))
+          .build();
+
+      // Define the instance properties for the template
+      InstanceProperties instanceProperties = InstanceProperties.newBuilder()
+          .addDisks(attachedDisk)
+          .setMachineType(machineType)
+          .addNetworkInterfaces(networkInterface)
+          .build();
+
+      // Build the instance template object
+      InstanceTemplate instanceTemplate = InstanceTemplate.newBuilder()
+          .setName(templateName)
+          .setProperties(instanceProperties)
+          .build();
+
+      // Create the request to insert the instance template
+      InsertRegionInstanceTemplateRequest insertInstanceTemplateRequest =
+          InsertRegionInstanceTemplateRequest
+              .newBuilder()
+              .setProject(projectId)
+              .setRegion(region)
+              .setInstanceTemplateResource(instanceTemplate)
+              .build();
+
+      // Send the request and wait for the operation to complete
+      Operation response = templatesClientRegion.insertAsync(insertInstanceTemplateRequest)
+          .get(3, TimeUnit.MINUTES);
+
+      if (response.hasError()) {
+        System.out.println("Instance Template creation failed! " + response);
+        return;
+      }
+      System.out.printf("Instance Template Operation Status: %s%n", response.getStatus());
+    }
+  }
+
+
+  // Delete an instance template with the REGIONAL location.
+  private static void deleteRegionalInstanceTemplate(
+      String projectId, String zone, String templateName)
+      throws IOException, ExecutionException, InterruptedException, TimeoutException {
+    try (RegionInstanceTemplatesClient regionInstanceTemplatesClient =
+             RegionInstanceTemplatesClient.create()) {
+      String region = zone.substring(0, zone.lastIndexOf('-')); // Extract the region from the zone
+
+      DeleteRegionInstanceTemplateRequest deleteInstanceTemplateRequest =
+          DeleteRegionInstanceTemplateRequest
+              .newBuilder()
+              .setProject(projectId)
+              .setRegion(region)
+              .setInstanceTemplate(templateName)
+              .build();
+
+      Operation response = regionInstanceTemplatesClient.deleteAsync(
+          deleteInstanceTemplateRequest).get(3, TimeUnit.MINUTES);
+
+      if (response.hasError()) {
+        System.out.println("Instance template deletion failed ! ! " + response);
+        return;
+      }
+      System.out.printf("Instance template deletion operation status for %s: %s ", templateName,
+          response.getStatus());
+    }
   }
 }
